feat: drag-and-drop support for StringFileUploadField
Add support for loading a local file into a StringFileUploadField via drag and drop.  Dragging a file onto the button will change the mouse cursor to indicate whether or not the file is acceptable according to the fileInputTypesAccepted filter, and dropping an acceptable file behaves the same as clicking the button and selecting that same file from the usual file chooser.

N.B. added dependency on https://www.npmjs.com/package/attr-accept to test dragged files against the same "accept" value as the regular file input.

Closes #576