fix CSS hover rules in AssistantTextSpanClassification (#987)
Hover rules in a couple of `sx` properties were mistakenly entered as `:hover`, meaning that once the relevant CSS fragment was added to the DOM, _every_ element in the whole document gained the green-background hover behaviour, not just the element corresponding to this component.

Fixed by changing them to `&:hover` to target this component specifically.