assistant: remove "abstract" from NE cloud tooltip (#997)
* fix: remove abstract from SPARQL query

Something has changed in dbpedia and the query we use to determine named entity types no longer works if we ask for the `rdfs:comment`.  For now, changed to remove this from the query so at least we still get the types right, at a later stage we can see about re-instating the full tooltip.

* fix: remove variable from SPARQL SELECT